PI/requirements.txt
jupyter
pandas
numpy
python-dotenv
mysql-connector-python
